@@ -6,8 +6,14 @@ from setuptools import setup, find_packages
 from os import path
 from io import open
 
+__author__ = "Luca Ponzoni"
+__date__ = "December 2019"
+__maintainer__ = "Luca Ponzoni"
+__email__ = "lponzoni@pitt.edu"
+__status__ = "Production"
+
 here = path.abspath(path.dirname(__file__))
-with open(path.join(here, 'VERSION'), encoding='utf-8') as f:
+with open(path.join(here, 'rhapsody/VERSION'), encoding='utf-8') as f:
     version = f.read()
 with open(path.join(here, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
@@ -16,19 +22,22 @@ setup(
     name='prody-rhapsody',
     version=version,
     packages=find_packages(exclude=['contrib', 'docs', 'tests']),
-    python_requires='>=3.5, <4',
-    install_requires=['requests', 'numpy', 'scikit-learn', 'matplotlib',
-                      'biopython', 'pyparsing', 'prody'],
+    python_requires='>=3.6, <4',
+    install_requires=[
+        'requests', 'tqdm',
+        'numpy',  'scikit-learn', 'matplotlib',
+        'biopython', 'pyparsing', 'prody'],
     extras_require={
         'docs': ['sphinx >= 1.4', 'sphinx_theme', 'sphinx_rtd_theme']
     },
+    include_package_data=True,
     description="""Python program, based on ProDy, for pathogenicity prediction
     of human missense variants.""",
     long_description=long_description,
     long_description_content_type='text/markdown',
     url='https://github.com/prody/rhapsody',
     author='Luca Ponzoni',
-    author_email='ponzoniluca@gmail.com',
+    author_email='lponzoni@pitt.edu',
     platforms=['Windows', 'MacOS X', 'POSIX'],
     license='GPL',
     classifiers=[
@@ -43,8 +52,6 @@ setup(
         'Operating System :: POSIX',
         'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
 
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
     ],
